+++ cellst/core/orchestrator.py
@@ -14,7 +14,7 @@ from cellst.core.arrays import ExperimentArray
 from cellst.utils._types import Arr
 from cellst.utils.process_utils import condense_operations, extract_operations
 from cellst.utils.log_utils import get_logger, get_console_logger
-from cellst.utils.yaml_utils import (save_operation_yaml, save_pipeline_yaml,
+from cellst.utils.file_utils import (save_operation_yaml, save_pipeline_yaml,
                                      save_yaml_file, folder_name)
 from cellst.utils.slurm_utils import JobController, SlurmController
 from cellst.utils.cli_utils import CLIParser
@@ -319,7 +319,7 @@ class Orchestrator():
             os.makedirs(path)
         path = os.path.join(path, fname)
 
-        # Save using yaml_utils
+        # Save using file_utils
         self.logger.info(f"Saving Operations at {path}")
         save_operation_yaml(path, self.operations)
 
@@ -329,7 +329,7 @@ class Orchestrator():
         if not os.path.exists(path):
             os.makedirs(path)
 
-        # Save using yaml_utils
+        # Save using file_utils
         self.logger.info(f"Saving condition_map at {path}")
         save_yaml_file(self.condition_map, path, warning=False)
