+++ cellst/utils/slurm_utils.py
@@ -10,7 +10,7 @@ from pprint import pprint
 from typing import Collection, Generator, List, Callable
 
 from cellst.utils.log_utils import get_console_logger, get_null_logger
-from cellst.utils.yaml_utils import save_job_history_yaml, get_file_line, load_yaml
+from cellst.utils.file_utils import save_job_history_yaml, get_file_line, load_yaml
 from cellst.core.pipeline import Pipeline
